requirements.txt
pymongo==4.4.0
streamlit
python-dotenv

